+++ src/include/ipc/probe.h
@@ -3,6 +3,8 @@
  * Copyright(c) 2022 Intel Corporation. All rights reserved.
  */
 
+#include <ipc/probe_dma_frame.h>
+
 #if CONFIG_IPC_MAJOR_4
 #include <ipc4/probe.h>
 #else
